clocks/requirements.txt
Django~=5.1.1
asgiref~=3.8.1
channels~=4.1.0
djangorestframework~=3.15.2
redis~=5.0.8
django-grappelli~=4.0.1
uvicorn
daphne==2.5.0
channels_redis
